+++ src/Symfony/Component/AssetMapper/CHANGELOG.md
@@ -6,6 +6,7 @@ CHANGELOG
 
  * Mark the component as non experimental
  * Add a `importmap:install` command to download all missing downloaded packages
+ * Allow specifying packages to update for the `importmap:update` command
 
 6.3
 ---